+++ src/aac_metrics/classes/meteor.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from typing import Union
+from typing import Optional, Union
 
 from torch import Tensor
 
@@ -32,6 +32,7 @@ class METEOR(AACMetric[Union[tuple[dict[str, Tensor], dict[str, Tensor]], Tensor
         java_path: str = ...,
         java_max_memory: str = "2G",
         language: str = "en",
+        use_shell: Optional[bool] = None,
         verbose: int = 0,
     ) -> None:
         super().__init__()
@@ -40,6 +41,7 @@ class METEOR(AACMetric[Union[tuple[dict[str, Tensor], dict[str, Tensor]], Tensor
         self._java_path = java_path
         self._java_max_memory = java_max_memory
         self._language = language
+        self._use_shell = use_shell
         self._verbose = verbose
 
         self._candidates = []
@@ -47,14 +49,15 @@ class METEOR(AACMetric[Union[tuple[dict[str, Tensor], dict[str, Tensor]], Tensor
 
     def compute(self) -> Union[tuple[dict[str, Tensor], dict[str, Tensor]], Tensor]:
         return meteor(
-            self._candidates,
-            self._mult_references,
-            self._return_all_scores,
-            self._cache_path,
-            self._java_path,
-            self._java_max_memory,
-            self._language,
-            self._verbose,
+            candidates=self._candidates,
+            mult_references=self._mult_references,
+            return_all_scores=self._return_all_scores,
+            cache_path=self._cache_path,
+            java_path=self._java_path,
+            java_max_memory=self._java_max_memory,
+            language=self._language,
+            use_shell=self._use_shell,
+            verbose=self._verbose,
         )
 
     def extra_repr(self) -> str: