+++ src/aac_metrics/classes/spider_fl.py
@@ -88,9 +88,9 @@ class SPIDErFL(AACMetric[Union[tuple[dict[str, Tensor], dict[str, Tensor]], Tens
 
     def compute(self) -> Union[tuple[dict[str, Tensor], dict[str, Tensor]], Tensor]:
         return spider_fl(
-            self._candidates,
-            self._mult_references,
-            self._return_all_scores,
+            candidates=self._candidates,
+            mult_references=self._mult_references,
+            return_all_scores=self._return_all_scores,
             # CIDEr args
             n=self._n,
             sigma=self._sigma,